+++ type_infer/__init__.py
@@ -3,13 +3,10 @@ from type_infer import dtype
 from type_infer import api
 from type_infer import helpers
 
-from type_infer.api import ENGINES
-
 __version__ = '0.0.18'
 
 
 __all__ = [
     '__version__',
     'base', 'dtype', 'api', 'helpers',
-    'ENGINES'
 ]
