--- type_infer/bert/infer.py
@@ -1,1 +1,0 @@
-STABLE = False
